backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
cachetools==5.5.2
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==44.0.2
daphne==4.1.2
dj-database-url==2.3.0
dj-email-url==1.0.6
dj-rest-auth==7.0.1
Django==5.2.9
django-allauth==65.13.0
django-cache-url==3.4.5
django-celery-beat==2.8.0
django-cors-headers==4.9.0
django-filter==25.2
django-redis==5.4.0
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.29.0
environs==14.1.1
flower==2.0.1
google-api-core==2.24.2
google-api-python-client==2.164.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.69.1
httplib2==0.22.0
humanize==4.12.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.5.0
marshmallow==3.26.2
packaging==24.2
pillow==11.3.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.0.0
pyparsing==3.2.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.1
PyYAML==6.0.3
redis==5.2.1
referencing==0.37.0
requests==2.32.4
rpds-py==0.30.0
rsa==4.9
service-identity==24.2.0
setuptools==78.1.1
six==1.17.0
sqlparse==0.5.3
tornado==6.5
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.2.13
zope.interface==7.2
